Fix issue where zsh_history gets truncated
The root cause is the SAVEHIST variable. When I was testing my zsh
start up without oh-my-zsh, my history file reduced down to 1000 lines,
which has caused immense pain. I figured it it's because SAVEHIST
defaults to 1000, but oh-my-zsh sets it to 10000. So using default zsh
however briefly caused me to lose a bunch of history.

This fix creates 2 safeguards:
- always sets SAVEHIST to 50000 regardless of oh-my-zsh
- moves my history file to a non-default location